Move make-array to @ember/array
