+++ packages/ember/index.ts
@@ -73,7 +73,7 @@ import DataAdapter from '@ember/debug/data-adapter';
 import EmberError from '@ember/error';
 import { run } from '@ember/runloop';
 import { getOnerror, setOnerror } from '@ember/-internals/error-handling';
-import EmberArray, { A, NativeArray, isArray } from '@ember/array';
+import EmberArray, { A, NativeArray, isArray, makeArray } from '@ember/array';
 import MutableArray from '@ember/array/mutable';
 import ArrayProxy from '@ember/array/proxy';
 import Application, { getOwner, setOwner, onLoad, runLoadHooks } from '@ember/application';
@@ -206,7 +206,6 @@ const PartialEmber = {
   inspect: utils.inspect,
   generateGuid: utils.generateGuid,
   guidFor: utils.guidFor,
-  makeArray: utils.makeArray,
   uuid: utils.uuid,
   wrap: utils.wrap,
 
@@ -228,6 +227,7 @@ const PartialEmber = {
   Array: EmberArray,
   NativeArray,
   isArray,
+  makeArray,
 
   // ****@ember/array/mutable****
   MutableArray,
