+++ types/publish.mjs
@@ -168,7 +168,6 @@ const PREVIEW_MODULES = [
   '@ember/-internals/utils/lib/invoke.d.ts',
   '@ember/-internals/utils/lib/is_proxy.d.ts',
   '@ember/-internals/utils/lib/lookup-descriptor.d.ts',
-  '@ember/-internals/utils/lib/make-array.d.ts',
   '@ember/-internals/utils/lib/mandatory-setter.d.ts',
   '@ember/-internals/utils/lib/name.d.ts',
   '@ember/-internals/utils/lib/spec.d.ts',
@@ -203,6 +202,7 @@ const PREVIEW_MODULES = [
   '@ember/array/mutable.d.ts',
   '@ember/array/proxy.d.ts',
   '@ember/array/-internals.d.ts',
+  '@ember/array/lib/make-array.d.ts',
   '@ember/canary-features/index.d.ts',
   '@ember/component/helper.d.ts',
   '@ember/component/index.d.ts',
